+++ next.config.ts
@@ -7,10 +7,15 @@ const withPWA = require("next-pwa")({
   skipWaiting: true,
 });
 
+import packageJson from './package.json';
+
 const nextConfig: NextConfig = {
   output: 'export',
   images: {
     unoptimized: true,
+  },
+  env: {
+    NEXT_PUBLIC_APP_VERSION: packageJson.version,
   },
 };